Version 4.3.0,  Revision 15234,  Hash 02fd57f
MODIFIED *  .gitignore config/Makefile.in config/acx_report.m4 config/configure.ac config/report.in config/version.m4 config/version.m4_gpl config/yambo_folders.m4 configure include/version.inc

NEW *  config/msg_ydb.in config/repos.list config/scripts.m4

Additions:
- Finalized the addition of YDB as external supplementary lib.
  It is activated by using the --enable-ydb flag and a post-configure message is given.

Patch sent by:  marini <andrea.marini@cnr.it>

+++ config/yambo_folders.m4
@@ -1,4 +1,26 @@
 #
+#        Copyright (C) 2000-2018 the YAMBO team
+#              http://www.yambo-code.org
+#
+# Authors (see AUTHORS file for details): DS AM
+#
+# This file is distributed under the terms of the GNU
+# General Public License. You can redistribute it and/or
+# modify it under the terms of the GNU General Public
+# License as published by the Free Software Foundation;
+# either version 2, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will
+# be useful, but WITHOUT ANY WARRANTY; without even the
+# implied warranty of MERCHANTABILITY or FITNESS FOR A
+# PARTICULAR PURPOSE.  See the GNU General Public License
+# for more details.
+#
+# You should have received a copy of the GNU General Public
+# License along with this program; if not, write to the Free
+# Software Foundation, Inc., 59 Temple Place - Suite 330,Boston,
+# MA 02111-1307, USA or visit http://www.gnu.org/copyleft/gpl.txt.
+#
 chmod u+x sbin/*
 cp config/Makefile .
 #
@@ -91,11 +113,17 @@ if ! test -d "$extlibs_path/${FCKIND}/${FC}/include"; then mkdir    "$extlibs_pa
 if ! test -d "$extlibs_path/${FCKIND}/${FC}/bin";     then mkdir    "$extlibs_path/${FCKIND}/${FC}/bin";     fi
 #
 if  test -d "$extlibs_path/${FCKIND}/${FC}" ; then
- for file in $extlibs_path/${FCKIND}/${FC}/bin/*; do
-  cp $file $exec_prefix/bin/
+ FILES=$extlibs_path/${FCKIND}/${FC}/bin/*
+ for file in $FILES; do
+  if test -f $file; then 
+   cp $file $exec_prefix/bin/
+  fi
  done
- for bin in $extlibs_path/${FCKIND}/${FC}/v3/bin/*; do
-  cp $bin $exec_prefix/bin/
+ FILES=$extlibs_path/${FCKIND}/${FC}/v3/bin/*
+ for file in $FILES; do
+  if test -f $file; then 
+   cp $file $exec_prefix/bin/
+  fi
  done
 fi
 #
